+++ resources/views/layout/target/show.blade.php
@@ -31,4 +31,5 @@ $(function () {
     $('[data-toggle="tooltip"]').tooltip()
 })
 </script>
+
 @endpush