+++ tests/Feature/TargetTest.php
@@ -1,0 +1,31 @@
+<?php
+/**
+ * Tests for creating, deleting, and adapting targets.
+ *
+ * PHP Version 7
+ *
+ * @category Test
+ * @package  DeepskyLog
+ * @author   Wim De Meester <deepskywim@gmail.com>
+ * @license  GPL3 <https://opensource.org/licenses/GPL-3.0>
+ * @link     http://www.deepskylog.org
+ */
+
+namespace Tests\Feature;
+
+use Tests\TestCase;
+use Illuminate\Foundation\Testing\RefreshDatabase;
+
+/**
+ * Tests for creating, deleting, and adapting targets.
+ *
+ * @category Test
+ * @package  DeepskyLog
+ * @author   Wim De Meester <deepskywim@gmail.com>
+ * @license  GPL3 <https://opensource.org/licenses/GPL-3.0>
+ * @link     http://www.deepskylog.org
+ */
+class TargetTest extends TestCase
+{
+    use RefreshDatabase;
+}